Revert "Merge pull request #22 from OnGle/migrate-to-phoenix-cart"
This reverts commit a0e74438b540c4169de3fbfc2be322828d45b7c7, reversing
changes made to d910bb6e629a5d32a3ea50cc066cab49953f83d5.

The original plan was to rename the oscommerce appliance build code
repo - to phoenixcart. However after some consideration we've decided
to create a new phoenixcart repo. It will include all the oscommerce
git history - as that is it's history/base but phoenixcart should be
considered a new appliance. So this PR is being reverted and the
oscommerce repo will be archived.